ragflow_demo/requirements.txt
# RAGFlow Python SDK - Official installation method:
# pip install ragflow-sdk

# For latest development version:
# pip install git+https://github.com/ragflow/ragflow-sdk.git

# Core dependencies:
pyyaml>=6.0
requests>=2.28.0
python-dotenv>=1.0.0
click>=8.0.0
tqdm>=4.60.0
colorama>=0.4.0

# RAGFlow SDK (optional - will be installed if needed)
# ragflow-sdk>=0.1.3